DjangoRestApi/tutorials/serializers.py
from django.db.models import fields
from rest_framework import serializers 
from tutorials.models import Tutorial, Users
 
 
class TutorialSerializer(serializers.ModelSerializer):
 
    class Meta:
        model = Tutorial
        fields = ('id',
                  'title',
                  'description',
                  'published')

class UsersSerializer(serializers.ModelSerializer):
    class Meta:
        model = Users
        fields = "__all__"
