DjangoRestApi/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
docopt==0.6.2
filelock==3.3.2
idna==3.3
platformdirs==2.4.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
yarg==0.1.9
